pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dolos"
version = "0.1.0"
description = "Word document manipulation tool with granular metadata control for creating, editing, and sanitizing document histories"
authors = [{name = "Dolos Team"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "python-docx>=1.1.0",
    "lxml>=5.0.0",
    "typer[all]>=0.12.0",
    "python-dateutil>=2.8.0",
    "sqlalchemy>=2.0.0",
    "nltk>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
dolos = "dolos.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
